Jira and Linear APi updates Code review
Category of a Linear workflow state https://linear.app/docs/configuring-workflow

+++ src/main/java/bogdanpc/linearsync/linear/control/AttachmentConfig.java
@@ -20,6 +20,6 @@ public interface AttachmentConfig {
 
         @WithName("max-size")
         @WithDefault("10485760")
-        long maxSize();
+        int maxSize();
     }
 }
